refactor: move pages to root
