+++ pages/api/auth/[...auth0].ts
