+++ pages/api/database/persons/count.ts
