+++ pages/api/database/properties/index.ts
