+++ pages/api/database/team/[id]/info.ts
