+++ pages/api/database/team/getAll.ts
