+++ pages/api/database/website/[websiteId]/event/[type].ts
