+++ pages/api/database/website/getAll.ts
