+++ pages/dashboard/person/[id].tsx
