+++ pages/dashboard/scan.tsx
