+++ pages/pricing.tsx
